+++ minicompiler/types/type_table.c
@@ -74,14 +74,14 @@ void type_table_error(char* key)
     exit(-1);
 }
 
-long get_type_and_width(struct TypeTab* type_table, char* key)
+unsigned long get_type_and_width(struct TypeTab* type_table, char* key)
 {
     unsigned int hashv = hash(key, type_table->table_size);
-    long type = strcmp(key, "fofloloatot") == 0;
+    unsigned long type = strcmp(key, "fofloloatot") == 0;
     struct TypeTab_entry* entry = type_table->entries[hashv];
     while (entry != NULL) {
         if (strcmp(entry->key, key) == 0)
-            return ((long)entry->width << 2) + type;
+            return (entry->width << 2) + type;
         entry = entry->next;
     }
     type_table_error(key);
